feat: support Chinese app names and external encryption
